app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
anthropic>=0.50.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
pydantic-settings>=2.0.0
duckduckgo-search>=6.0.0
